:recycle: [Fix] 회원가입 URI 변경 및 메소드위치를 MemberApiController로 옮김
- URI 변경 : /api/v1/register -> /api/v1/members
- swagger 주석 추가
