+++ src/main/java/com/karrotclone/api/MemberApiController.java
@@ -3,6 +3,7 @@ package com.karrotclone.api;
 import com.karrotclone.domain.Favorite;
 import com.karrotclone.domain.Member;
 import com.karrotclone.domain.SalesPost;
+import com.karrotclone.dto.RegisterDto;
 import com.karrotclone.dto.ResponseDto;
 import com.karrotclone.dto.SalesPostDetailDto;
 import com.karrotclone.dto.SalesPostSimpleDto;
@@ -10,12 +11,16 @@ import com.karrotclone.exception.DomainNotFoundException;
 import com.karrotclone.repository.FavoriteRepository;
 import com.karrotclone.repository.SalesPostRepository;
 import com.karrotclone.repository.TempMemberRepository;
+import com.karrotclone.service.auth.RegisterService;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -35,6 +40,28 @@ public class MemberApiController {
     private final SalesPostRepository salesPostRepository;
     private final TempMemberRepository memberRepository;
     private final FavoriteRepository favoriteRepository;
+    private final RegisterService registerService;
+
+    @ApiOperation(value="회원가입 요청", notes="입력된 정보를 토대로 회원가입을 진행합니다.")
+    @PostMapping("/api/v1/members")
+    public void register(final @Valid @RequestBody RegisterDto dto, BindingResult bindingResult){
+        if (bindingResult.hasErrors()){
+            System.out.println(" Incoming error ");
+            return;
+        }
+        try {
+            registerService.register(dto);
+            System.out.println("회원 가입이 완료되었습니다.");
+        }catch (DataIntegrityViolationException e){
+            e.printStackTrace();
+            bindingResult.reject("register failed","이미 등록된 메일입니다.");
+        } catch (Exception e){
+            System.out.println("error = " + e);
+            e.printStackTrace();
+            bindingResult.reject("register failed", e.getMessage());
+        }
+    } //추후에 다른 로그인 Auth 구현 가능.
+
 
     /**
      * 거래글을 관심목록에 추가하거나, 이미 관심목록에 있다면 삭제합니다.
